split into mobile and desktop daycard
