+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './Header';
 export * from './DayCard';
 export { ThemeProvider, useTheme } from './ThemeContext';
+export * from './DesktopFeaturedCard';
